+++ src/components/Documents/DocumentTable.jsx
@@ -16,14 +16,19 @@ import DocumentTableRow from './DocumentTableRow';
 import { EmptyListNotification, LoadingAnimation } from '../Notification';
 
 /**
+ * @typedef {import("../../typedefs.js").documentTableProps} documentTableProps
+ */
+
+/**
  * DocumentTable Component - The Document Table that shows the list of documents
  * stored on Solid
  *
  * @memberof Documents
  * @name DocumentTable
+ * @param {documentTableProps} Props - Props for DocumentTable component
  * @returns {React.JSX.Element} The DocumentTable component
  */
-const DocumentTable = () => {
+const DocumentTable = ({ handleAclPermissionsModal }) => {
   const { documentListObject, loadingDocuments } = useContext(DocumentListContext);
   const columnTitlesArray = [
     'Name',
@@ -32,6 +37,7 @@ const DocumentTable = () => {
     'Upload Date',
     'Expiration Date',
     'Preview File',
+    'Permissions',
     'Delete'
   ];
 
@@ -51,7 +57,11 @@ const DocumentTable = () => {
           </TableHead>
           <TableBody>
             {documentListObject?.docList.map((document) => (
-              <DocumentTableRow key={document.name} document={document} />
+              <DocumentTableRow
+                key={document.name}
+                document={document}
+                handleAclPermissionsModal={handleAclPermissionsModal}
+              />
             ))}
           </TableBody>
         </Table>
