+++ src/components/Form/index.js
@@ -1,7 +1,5 @@
 import DocumentSelection from './DocumentSelection';
 import FormSection from './FormSection';
-import SetAclPermissionForm from './SetAclPermissionForm';
-import SetAclPermsDocContainerForm from './SetAclPermsDocContainerForm';
 /**
  * Components and functions related to forms within project PASS and used for
  * the Documents Page
@@ -9,4 +7,4 @@ import SetAclPermsDocContainerForm from './SetAclPermsDocContainerForm';
  * @namespace Forms
  */
 
-export { DocumentSelection, FormSection, SetAclPermissionForm, SetAclPermsDocContainerForm };
+export { DocumentSelection, FormSection };
