+++ src/components/Modals/index.js
@@ -2,6 +2,7 @@ import AddContactModal from './AddContactModal';
 import DeleteContactModal from './DeleteContactModal';
 import LogoutModal from './LogoutModal';
 import NewMessageModal from './NewMessageModal';
+import SetAclPermissionsModal from './SetAclPermissionsModal';
 import UploadDocumentModal from './UploadDocumentModal';
 
 /**
@@ -10,4 +11,11 @@ import UploadDocumentModal from './UploadDocumentModal';
  * @namespace Modals
  */
 
-export { AddContactModal, DeleteContactModal, LogoutModal, NewMessageModal, UploadDocumentModal };
+export {
+  AddContactModal,
+  DeleteContactModal,
+  LogoutModal,
+  NewMessageModal,
+  SetAclPermissionsModal,
+  UploadDocumentModal
+};
